+++ docs/make.jl
@@ -48,6 +48,7 @@ makedocs(;
             "vnt/arraylikeblocks.md",
         ],
         "Initialisation strategies" => "init.md",
+        "Link strategies" => "link.md",
         "Accumulators" => "accumulators.md",
         "Model evaluation" => "flow.md",
         "Storing values" => "values.md",