+++ src/onlyaccs.jl
@@ -24,28 +24,18 @@ end
 DynamicPPL.maybe_invlink_before_eval!!(vi::OnlyAccsVarInfo, ::Model) = vi
 DynamicPPL.getaccs(vi::OnlyAccsVarInfo) = vi.accs
 DynamicPPL.setaccs!!(::OnlyAccsVarInfo, accs::AccumulatorTuple) = OnlyAccsVarInfo(accs)
+function DynamicPPL.get_transform_strategy(::OnlyAccsVarInfo)
+    # OAVI doesn't contain this info, we can't return a sensible value. Hopefully this
+    # method doesn't ever get called though.
+    return error(
+        "get_transform_strategy cannot be implemented for OnlyAccsVarInfo; please specify a transform strategy manually in your call to `init!!`",
+    )
+end
 
-# Ideally, we'd define this together with InitContext, but alas that file comes way before
-# this one, and sorting out the include order is a pain.
-function tilde_assume!!(
-    ctx::InitContext,
-    dist::Distribution,
-    vn::VarName,
-    template::Any,
-    vi::Union{OnlyAccsVarInfo,ThreadSafeVarInfo{<:OnlyAccsVarInfo}},
+# This allows us to make use of the main tilde_assume!!(::InitContext) method without
+# having to duplicate the code here
+@inline function DynamicPPL.setindex_with_dist!!(
+    vi::OnlyAccsVarInfo, ::AbstractTransformedValue, ::Distribution, ::VarName, ::Any
 )
-    # For OnlyAccsVarInfo, since we don't need to write into the VarInfo, we can 
-    # cut out a lot of the code above.
-    tval = init(ctx.rng, vn, dist, ctx.strategy)
-    # Prefer to use the transform from the distribution.
-    transform = if tval isa LinkedVectorValue
-        DynamicPPL.from_linked_vec_transform(dist)
-    elseif tval isa VectorValue
-        DynamicPPL.from_vec_transform(dist)
-    else
-        DynamicPPL.get_transform(tval)
-    end
-    x, inv_logjac = with_logabsdet_jacobian(transform, DynamicPPL.get_internal_value(tval))
-    vi = accumulate_assume!!(vi, x, tval, -inv_logjac, vn, dist, template)
-    return x, vi
+    return vi
 end